+++ packages/expo/ios/AppDelegates/ExpoReactNativeFactoryDelegate.swift
@@ -4,7 +4,7 @@ open class ExpoReactNativeFactoryDelegate: RCTDefaultReactNativeFactoryDelegate 
   open override func customize(_ rootView: UIView) {
     ExpoAppDelegateSubscriberRepository.subscribers.forEach { $0.customizeRootView?(rootView) }
   }
-  
+
   open override func createRootViewController() -> UIViewController {
     return ExpoAppDelegateSubscriberRepository.reactDelegateHandlers.lazy
       .compactMap { $0.createRootViewController() }